+++ DTO/RevenueByDateDTO.cs
@@ -3,6 +3,6 @@
     public class RevenueByDateDTO
     {
         public DateTime Date { get; set; }
-        public decimal Revenue { get; set; }
+        public double Revenue { get; set; }
     }
 }
